code/requirements.txt
facenet-pytorch==2.6.0
pillow==10.2.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
gradio==5.3.0
pinecone==5.3.1
requests==2.32.3
pyngrok==7.2.0
python-dotenv==1.0.1